% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area under the curve}
\usage{
auc(x, ...)
}
\arguments{
\item{x}{object of class \code{roc}.}

\item{...}{other arguments (none are used at the moment).}
}
\value{
returns the area under the curve for a \code{roc} class object.
}
\description{
Calculate the area under the curve

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
a <- roc(QIDS$QIDS, QIDS$depression, c("Yes","No"), "Yes")
auc(a)
}
\author{
Mathijs Deen
}
