% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check model for influential cases}
\usage{
check(object, ...)
}
\arguments{
\item{object}{object of class \code{lm}.}

\item{...}{other parameters (none are used at the moment).}
}
\value{
\code{check} returns a list containing a matrix with statistics regarding
influential cases and a vector of variance inflation factors. Furthermore, it
produces diagnostics plots.\cr
The return list contains two elements:\cr \cr
- \code{influence}, a \code{data.frame}, with observations in the model,
and the following variables:
\item{predicted.value}{The value predicted by the model.}
\item{residual}{The raw residual.}
\item{std.residual}{The standardized residual.}
\item{dfb.<...>}{DFBETAs for the variables in the model.}
\item{dffit}{DFFIT value.}
\item{cov.r}{Covariance ratio, a measure of change in the determinant of
the coefficient covariance matrix.}
\item{cook.d}{Cook's distance.}
\item{hat}{Hat values.}
\item{influential}{Determines whether a case is influential on any of the
measures \code{dfb.<...>}, \code{dffit}, \code{cov.r}, \code{cook.d} or \code{hat}.
See \code{influential cases} for more information.}
\cr
- \code{vifs}, a vector containing variance inflation factors for the
variables in the model.
\cr
The generated plots are the plots produced by \code{plot.lm}, numbers 1 through 6.
}
\description{
Check model for influential cases
}
\section{influential cases}{


For the influence indicators, the following rules are applied to check whether a case
is influential:
\itemize{
\item \eqn{\mathrm{any\enspace}|\mathrm{dfbeta}| > 1}.
\item \eqn{|\mathrm{dffit}| > 3 \sqrt{\frac{k}{n-k}}}.
\item \eqn{|1 - \mathrm{cov.r}| > \frac{3k}{n-k}}.
\item \eqn{F\mathrm{(}n, n-k \mathrm{)} = \mathrm{cooks.d\enspace having\enspace}.
  p > .5}
\item \eqn{\mathrm{hat} > \frac{3k}{n}}.
}
These indicators for being an influential case were derived from
\code{\link[stats]{influence.measures}} in the \code{stats} package.
}

\examples{
lm.1 <- lm(mpg ~ disp + wt, data = mtcars)
check(lm.1)
}
\author{
Mathijs Deen
}
