% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.range.R
\name{\%in.range\%}
\alias{\%in.range\%}
\title{Return which values are in a certain range}
\usage{
lhs \%in.range\% rhs
}
\arguments{
\item{lhs}{numeric vector}

\item{rhs}{numeric vector of length 2 with the bounds of the range}
}
\value{
\code{\%in.range\%} returns a logical vector of \code{length(lhs)}, indicating which
values of \code{lhs} are and are not in range \code{rhs}. Boundaries of \code{rhs} are included.
}
\description{
\code{\%in.range\%} indicates which values are in a certain range, including the
boundaries of the range.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
a <- seq(0, 1, 0.05)
r <- c(0.4,0.7)
cbind(a,
      'a \%in.range\% r' = a \%in.range\% r,
      'a \%within.range\% r' = a \%within.range\% r)
}
\seealso{
\code{\link{\%within.range\%}}
}
\author{
Mathijs Deen
}
