% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni.R
\name{\%ni\%}
\alias{\%ni\%}
\title{\%ni\% function}
\usage{
lhs \%ni\% rhs
}
\arguments{
\item{lhs}{left hand side}

\item{rhs}{right hand side}
}
\value{
The function returns a vector of the same length as \code{lhs}.
}
\description{
Evaluates whether the left hand side argument is not in the right hand side argument.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
The \code{\%ni\%} function negates the \code{\%in\%} operator.
}
\examples{
c(1,2,3) \%ni\% c(1,2)
}
\author{
Mathijs Deen
}
