% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{latinSquare}
\alias{latinSquare}
\title{Latin square}
\usage{
latinSquare(n)
}
\arguments{
\item{n}{positive integer}
}
\value{
\code{latinSquare} returns a matrix of size \eqn{n \times n} that is a latin square.
}
\description{
Create a latin square of size \eqn{n \times n}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
The latin square that is produced is a (generalized) latin square that is also balanced for first order carryover effects.
}
\examples{
# 5 x 5 latin square
latinSquare(5)
}
\references{
\itemize{
\item Williams, E.J. (1949). Experimental designs balanced for the estimation of residual effects of treatments. \emph{Australian Journal of Scientific Research 2, 2}, 149-168. https://doi.org/10.1071/CH9490149
}
}
\author{
Mathijs Deen
}
