% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mC1}
\alias{mC1}
\title{Center columns of a numeric matrix}
\usage{
mC1(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
A centered numeric matrix.
}
\description{
Center columns of a numeric matrix
}
\examples{
m <- matrix(c(1, 2, NA, 4, 5, 6), nrow = 3)
mC1(m)
}
