% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdma_session.R
\name{mdma_session}
\alias{mdma_session}
\title{Launch the MDMA Shiny application}
\usage{
mdma_session(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{\verb{[logical(1)]} Open a browser window automatically.
Defaults to \code{TRUE}.}

\item{...}{Passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
Called for its side effect.
}
\description{
Opens an interactive Shiny application for running \code{\link[=mdma_pdf]{mdma_pdf()}} and
\code{\link[=mdma_clean]{mdma_clean()}}.
}
