% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEAC.R
\name{plot.CEAC}
\alias{plot.CEAC}
\title{Plot cost-effectiveness acceptability curve}
\usage{
\method{plot}{CEAC}(
  x,
  xlab = "Cost-effectiveness threshold (€)",
  ylab = "Probability that intervention is cost-effective",
  las = 1,
  xlim = c(0, max(x$s$ICERs)),
  ...
)
}
\arguments{
\item{x}{object of class \code{CEAC}}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{las}{style of the axis labels (see \code{\link[graphics]{par}})}

\item{xlim}{limits of the x axis}

\item{...}{other arguments to be passed to the \code{plot} function.}
}
\value{
returns a plot
}
\description{
Plot cost-effectiveness acceptability curve
}
\examples{
\dontrun{1}
}
\author{
Mathijs Deen
}
