% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.probeInteraction.R
\name{plot.probeInteraction}
\alias{plot.probeInteraction}
\title{plot probed interaction}
\usage{
\method{plot}{probeInteraction}(
  x,
  ...,
  lty.p = "solid",
  pch.p = 1,
  col.p = "black",
  lty.effect = "longdash",
  pch.effect = 17,
  col.effect = "green",
  col.JN = "red",
  lty.JN = "dotted",
  pos.legend = "topleft"
)
}
\arguments{
\item{x}{object of class \code{probeInteraction}}

\item{...}{other arguments (none are used)}

\item{lty.p}{linetype for p values}

\item{pch.p}{point character for p values}

\item{col.p}{color for p values}

\item{lty.effect}{linetype for effect}

\item{pch.effect}{point character for effect}

\item{col.effect}{color for effect}

\item{col.JN}{color for Johnson-Neyman cut-off line(s)}

\item{lty.JN}{linetype for Johnson-Neyman cut-off line(s)}

\item{pos.legend}{position of legend}
}
\value{
\code{plot.probeInteraction} returns a combined plot with p value on the
first y axis and effect of the antecedent variable
}
\description{
Plot the effects of the antecedent and corresponding p value
as functions of the moderator.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
lm.1 <- lm(mpg ~ hp * wt, data = mtcars)
pI <- probeInteraction(lm.1, hp, wt, JN=TRUE, n.interval.moderator = 3,
                       quantile.moderator = c(0.1,0.9), values.moderator = 2)
plot(pI)
}
\author{
Mathijs Deen
}
