% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{print.check.lm}
\alias{print.check.lm}
\title{Print lm check}
\usage{
\method{print}{check.lm}(x, which.infl = c("influential", "all"), ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{which.infl}{Indicate whether only influential cases (\code{influential}, the default)
or all cases (\code{all}) should be printed.}

\item{...}{further arguments passed to or from other methods (none are used).}
}
\value{
prints the \code{check.lm} object.
}
\description{
Print the check of lm object

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
lm.1 <- lm(mpg ~ disp + wt, data = mtcars)
chk.lm.1 <- check(lm.1)
print(chk.lm.1, which.infl="all")
}
\author{
Mathijs Deen
}
