% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocMethods.R
\name{print.rocc}
\alias{print.rocc}
\title{Print roc object}
\usage{
\method{print}{rocc}(x, ...)
}
\arguments{
\item{x}{roc object}

\item{...}{other arguments}
}
\value{
prints an roc object
}
\description{
Print roc object
}
\examples{
\dontrun{
1}
}
\author{
Mathijs Deen
}
