% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTestMethods.R
\name{print.tTest}
\alias{print.tTest}
\title{Print t-test}
\usage{
\method{print}{tTest}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\value{
prints the \code{tTest} object as a \code{htest} object.
}
\description{
Print t-test
}
\examples{
x1 <- QIDS$QIDS[QIDS$depression == "Yes"]
x2 <- QIDS$QIDS[QIDS$depression == "No"]
tt <- tTest(x1, x2)
print(tt)
}
\author{
Mathijs Deen
}
