% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtp.R
\name{print.wtp}
\alias{print.wtp}
\title{Print willingness to pay probe}
\usage{
\method{print}{wtp}(x, ...)
}
\arguments{
\item{x}{object of class \code{wtp}.}

\item{...}{other arguments (none are used).}
}
\value{
\code{print.wtp} prints the outcome of \code{wtp}
}
\description{
Print the outcome of a willingness to pay threshold probe.
}
\examples{
CEA(gnomes, insulationMethod, Costs, diffHATS, 1000, "acorns") |>
  CEAC() |>
  wtp(probability = 0.80)
CEA(gnomes, insulationMethod, Costs, diffHATS, 1000, "acorns") |>
  CEAC() |>
  wtp(threshold = 8)
}
\author{
Mathijs Deen
}
