% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeInteraction.R
\name{probeInteraction}
\alias{probeInteraction}
\title{Probe interaction models}
\usage{
probeInteraction(
  object,
  antecedent,
  moderator,
  alpha = 0.05,
  JN = TRUE,
  n.interval.moderator,
  quantile.moderator,
  values.moderator
)
}
\arguments{
\item{object}{describe parameter}

\item{antecedent}{describe parameter}

\item{moderator}{describe parameter}

\item{alpha}{describe parameter}

\item{JN}{describe parameter}

\item{n.interval.moderator}{describe parameter}

\item{quantile.moderator}{describe parameter}

\item{values.moderator}{describe parameter}
}
\value{
\code{probeInteraction} returns a data frame containing values of the moderator
in a linear model, the effect of the antecedent at that value of the moderator,
standard errors, t values, p values and a confidence interval.
}
\description{
Probe the effect of a moderator on an X/antecedent variable in a linear model.
}
\examples{
lm.1 <- lm(mpg ~ hp * wt, data = mtcars)
probeInteraction(lm.1, hp, wt, JN=TRUE, n.interval.moderator = 3,
                 quantile.moderator = c(0.1,0.9), values.moderator = 2)
}
\author{
Mathijs Deen
}
