% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Receiver operator characteristic}
\usage{
roc(response, group, levels, state)
}
\arguments{
\item{response}{response variable}

\item{group}{group variable}

\item{levels}{level variable}

\item{state}{state level}
}
\value{
blaat
}
\description{
Receiver operator characteristic
}
\examples{
\dontrun{
x <- 1}
}
