% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\title{Simultaneously print and return an object}
\usage{
show(object)
}
\arguments{
\item{object}{the object that should be printed and returned.}
}
\value{
\code{show} returns the value of \code{object}, while also printing it.
}
\description{
\code{show} prints and returns an object. This is useful
if one wants to show intermediate results in a pipe sequence.
}
\examples{
mtcars |>
lm(mpg ~  disp + hp, data = _) |>
keep(lm.mpg_disp_hp) |>
summary() |>
show() |>
.Primitive("$")("df")
}
\author{
Mathijs Deen
}
