% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocMethods.R
\name{auc}
\alias{auc}
\alias{auc.roc}
\title{Area under the curve}
\usage{
auc(x)

\method{auc}{roc}(x)
}
\arguments{
\item{x}{object of class \code{roc}.}
}
\value{
returns the area under the curve for a \code{roc} class object.
}
\description{
Area under the curve
}
\examples{
a <- roc(QIDS$QIDS, QIDS$depression, c("Yes","No"), "Yes")
auc(a)
}
\author{
Mathijs Deen
}
