% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEA.R
\name{CEA}
\alias{CEA}
\title{Cost-effectiveness analysis}
\usage{
CEA(data, group, cost, effect, B = 5000)
}
\arguments{
\item{data}{a \code{data.frame} with rows representing observations in for
example a cost-effectiveness trial.}

\item{group}{group variable in \code{data}. Should contain two levels.}

\item{cost}{cost variable in \code{data}.}

\item{effect}{effect variable in \code{data}.}

\item{B}{number of bootstrap samples.}
}
\value{
\code{CEA} returns a list (class \code{CEA}) with the following elements:
\item{stats}{a \code{data.frame} containing the bootstrap statistics: estimates for
the difference in costs (\code{diffC}), the difference in effects(\code{diffE}),
and the ICER as a ratio of these two.}
\item{diff.C.true}{Observed difference in costs.}
\item{diff.E.true}{Observed difference in effects.}
\item{ICER.true}{Observed incremental cost-effectiveness ratio.}
}
\description{
Cost-effectiveness analysis
}
\examples{
\dontrun{
1
}
}
\author{
Mathijs Deen
}
