% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEAC.R
\name{CEAC}
\alias{CEAC}
\title{Cost-effectiveness acceptability curve}
\usage{
CEAC(x)
}
\arguments{
\item{x}{object of class \code{CEA}}
}
\value{
\code{CEAC} returns data that can be plotted using \code{plot.CEAC}.
}
\description{
Create data for cost-effectiveness acceptability curve

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
CEA(gnomes, insulationMethod, Costs, diffHATS, 1000, "acorns") |>
  CEAC() |>
  plot(xlim = c(0,200))
}
\author{
Mathijs Deen
}
