% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEAC.R
\name{CEAC}
\alias{CEAC}
\title{Create data for cost-effectiveness acceptability curve}
\usage{
CEAC(x)
}
\arguments{
\item{x}{object of class \code{CEA}}
}
\value{
\code{CEAC} returns data that can be plotted using \code{plot.CEAC}.
}
\description{
Create data for cost-effectiveness acceptability curve
}
\examples{
\dontrun{1}
}
\author{
Mathijs Deen
}
