% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check model for influential cases}
\usage{
check(object, ...)
}
\arguments{
\item{object}{object of class \code{lm}}

\item{...}{other parameters (none are used at the moment)}
}
\value{
\code{check} returns a list containing a matrix with statistics regarding
influential cases and a vector of variance inflation factors. Furthermore, it
produces diagnostics plots.
}
\description{
Check model for influential cases
}
\examples{
lm.1 <- lm(mpg ~ disp + wt, data = mtcars)
check(lm.1)
}
\author{
Mathijs Deen
}
