% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corList.R
\name{corList}
\alias{corList}
\title{List of correlation coefficients}
\usage{
corList(x, ...)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{...}{arguments passed to the \code{cor} function.}
}
\value{
\code{corList} returns a list of correlations
}
\description{
List all correlations in a correlation matrix without duplicates.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
mtcars[,c("mpg","disp", "hp", "drat", "wt", "qsec")] |>
  corList(method="spearman")
}
\author{
Mathijs Deen
}
