% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency2unicode.R
\name{currency2unicode}
\alias{currency2unicode}
\title{Retrieve Unicode character for a currency}
\usage{
currency2unicode(currency, ...)
}
\arguments{
\item{currency}{character string. Supported values are accounting sign, afghani,
armenian dram, austral, baht, bitcoin, boliviano, cent, cedi, currency, dollar,
dong, drachma, dutch guilder, euro, franc, georgian lari, german penny, hryvnia,
indian rupee, iranian rial, kip, lari, lira, livre tournois, manat, mark, new shekel,
pakistani rupee, peso, pound, quetzal, real, rial, ruble, shekel, spesmilo, syrian pound,
tenge, tugrik, turkish lira, won, yen, yuan.}

\item{...}{other options (for troubleshooting only).}
}
\value{
\code{currency2unicode} the unicode character for a given currency.
}
\description{
Retrieve Unicode character for a currency
}
\details{
The input is evaluated case insensitive. In case the input is not supported, the function
will fall back to the generic currency symbol (\code{"¤"}).
}
\examples{
currency2unicode("dollar")

cat(sprintf("\%s5 is all my mom allows me to spend.", currency2unicode("dollar")))

}
\author{
Mathijs Deen
}
