% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{divMod}
\alias{divMod}
\title{divMod}
\usage{
divMod(x, d, shortForm = TRUE)
}
\arguments{
\item{x}{numeric value or vector}

\item{d}{numeric value or vector}

\item{shortForm}{boolean (defaults to \code{TRUE}), indicating whether the output should contain \code{x} and \code{d}}
}
\value{
\code{divMod} returns an object of class \code{divMod} (which is essentially a \code{data.frame}). It consists of the following values/variables:
\item{quotient}{The quotient of the integer division \code{x}/\code{n}}
\item{remainder}{The remainder of the division above}
Dependent upon the argument \code{shortForm}, the following values are present:
\item{x}{The argument \code{x}}
\item{d}{The argument \code{d}. Returning \code{x} and \code{d} is especially useful with the use of recycling (see example below).}
}
\description{
Retrieve quotient and remainder from Euclidian division.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
\code{divMod} is equivalent to Python's \code{divmod}, but extends it with support for vector recycling.
}
\examples{
divMod(x = 15, d = 4)
divMod(x = c(15,23,42,58), d = c(4,7), shortForm = FALSE) #recycling

renderTime <- function(milliseconds) {
  out <- list("hours"=NA, "minutes"=NA, "seconds"=NA, "milliseconds"=NA)
  out[c("seconds","milliseconds")] <- divMod(milliseconds,1000)
  out[c("minutes","seconds")] <- divMod(out$seconds,60)
  out[c("hours","minutes")] <- divMod(out$minutes, 60)
  with(out, sprintf("\%02d:\%02d:\%02d.\%03d", hours,minutes,seconds,milliseconds))
  }

renderTime(1000)
renderTime(68000)
renderTime(3683121)

}
\author{
Mathijs Deen
}
