% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{divMod}
\alias{divMod}
\title{divmod}
\usage{
divMod(x, d, shortForm = TRUE)
}
\arguments{
\item{x}{numeric value or vector}

\item{d}{numeric value or vector}

\item{shortForm}{boolean (defaults to \code{TRUE}), indicating whether the output should contain \code{x} and \code{d}}
}
\value{
\code{divMod} returns an object of class \code{divMod} (which is essentially a \code{data.frame}.)
Dependent upon the argument \code{shortForm}, the values of \code{x} and \code{d} are returned as well, which is especially useful when
recycling is used when \code{x} and \code{d} are of different length.
}
\description{
Return quotient and remainder from integer division of the input.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
divMod(x = 15, d = 4)
divMod(x = c(15,23,42,58), d = c(4,7), shortForm = FALSE) #recycling

hh_mm_ss <- function(milliseconds) {
  seconds <- divMod(round(milliseconds/1000),60)
  minutes <- divMod(seconds$quotient, 60)
  hours   <- minutes$quotient
  seconds <- seconds$remainder
  minutes <- minutes$remainder
  sprintf("\%02d:\%02d:\%02d", hours,minutes,seconds)
}

hh_mm_ss(1000)
hh_mm_ss(68000)
hh_mm_ss(3680000)

}
\author{
Mathijs Deen
}
