% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2Local.R
\name{f2Local}
\alias{f2Local}
\title{Calculate local \eqn{f^2} for (generalized) linear models}
\usage{
f2Local(object, method, ...)
}
\arguments{
\item{object}{a model object (currently supported: \code{lm}, \code{glm}, \code{vglm}).}

\item{method}{method for calculation of \eqn{R^2}, which is needed for the calculation of \eqn{f^2}. See Details.}

\item{...}{currently not used}
}
\value{
\code{f2Local} returns a list containing \eqn{f^2} values for every parameter in a model.
}
\description{
Calculate local \eqn{f^2} for (generalized) linear models
}
\details{
The following methods can be specified:
\itemize{
\item \code{lm} objects: \code{r.squared} and \code{adj.r.squared} as extracted from the \code{lm} object.
\item \code{glm} objects: \code{mcfadden}, \code{nagelkerke}, \code{coxsnell}, \code{tjur} and \code{efron}, as implemented
in the \code{performance} package.
\item \code{vglm} objects: \code{mcfadden}, \code{nagelkerke}, \code{coxsnell}, \code{tjur} and \code{efron}, as implemented
in the \code{R2.vglm} function.
}

Note that for multinomial models, using \code{method="efron"} gives questionable with \code{glm} objects and
is not possible for \code{vglm} objects. For \code{glm} objects, \code{method=coxsnell} cannot be used when the
response is not binary.
}
\examples{
model <- lm(mpg ~ cyl + wt*drat, data = mtcars)
}
\author{
Mathijs Deen
}
