% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Display frequency table}
\usage{
frequencies(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
object of type \code{data.frame} containing frequencies, percentages and cumulative percentages
}
\description{
Display frequency table with percentages and cumulative percentages
}
\examples{
\dontrun{
frequencies(datasets::mtcars$carb)}
}
\author{
Mathijs Deen
}
