% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTest.R
\name{print.tTest}
\alias{print.tTest}
\title{Print t test}
\usage{
\method{print}{tTest}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
prints the \code{tTest} object as a \code{htest} object.
}
\description{
Print the output of a t test.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
x1 <- QIDS$QIDS[QIDS$depression == "Yes"]
x2 <- QIDS$QIDS[QIDS$depression == "No"]
tt <- tTest(x1, x2)
print(tt)
}
\author{
Mathijs Deen
}
