% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIDS-data.R
\docType{data}
\name{QIDS}
\alias{QIDS}
\title{QIDS depression data}
\format{
the following variables are available:
\itemize{
\item \code{QIDS}: QIDS-SR total score
\item \code{Depression}: an indicator whether the individual was diagnosed with a depression or not.
}
}
\usage{
QIDS
}
\description{
The \code{QIDS} dataframe consists of 100 observations of people that were
clinically diagnosed with a major depressive disorder and who filled out the QIDS-SR questionnaire. The data were simulated.
}
